feat/M2-7(Criar interfaces para o banco de dados) :recycle: Refatoração da estrutura de pacotes no módulo adaptador
* A estrutura de pacotes no módulo adaptador foi refatorada.
* As classes relacionadas à configuração web, manipulador e modelo foram movidas do pacote de serviço geral para subpacotes `Rest`.
* A refatoração melhora a clareza e a possibilidade de manutenção do código.
