--- src/main/java/diegosneves/github/hexagonal/HexagonalArchitectureApplication.java
@@ -1,13 +1,0 @@
-package diegosneves.github.hexagonal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HexagonalArchitectureApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HexagonalArchitectureApplication.class, args);
-    }
-
-}
