--- src/main/java/diegosneves/github/hexagonal/adapters/dto/ExceptionDTO.java
@@ -1,4 +1,0 @@
-package diegosneves.github.hexagonal.adapters.dto;
-
-public record ExceptionDTO(String message, int statusCode) {
-}
