--- src/main/java/diegosneves/github/hexagonal/adapters/mapper/MapperStrategy.java
@@ -1,7 +1,0 @@
-package diegosneves.github.hexagonal.adapters.mapper;
-
-public interface MapperStrategy <T, E> {
-
-    T mapper(E origin);
-
-}
