--- src/main/java/diegosneves/github/hexagonal/adapters/request/ProductRequest.java
@@ -1,17 +1,0 @@
-package diegosneves.github.hexagonal.adapters.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-public class ProductRequest {
-    private String productName;
-    private Double productPrice;
-}
