--- src/main/java/diegosneves/github/hexagonal/adapters/response/ProductResponse.java
@@ -1,22 +1,0 @@
-package diegosneves.github.hexagonal.adapters.response;
-
-import diegosneves.github.hexagonal.app.enums.ProductStatus;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-public class ProductResponse {
-
-    private String id;
-    private String productName;
-    private ProductStatus status;
-    private Double productPrice;
-
-}
