+++ src/main/java/diegosneves/github/hexagonal/adapters/rest/HexagonalArchitectureApplication.java
@@ -1,0 +1,13 @@
+package diegosneves.github.hexagonal.adapters.rest;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HexagonalArchitectureApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HexagonalArchitectureApplication.class, args);
+    }
+
+}
