+++ src/main/java/diegosneves/github/hexagonal/adapters/rest/config/handler/ControllerExceptionHandler.java
@@ -1,0 +1,32 @@
+package diegosneves.github.hexagonal.adapters.rest.config.handler;
+
+import diegosneves.github.hexagonal.adapters.rest.dto.ExceptionDTO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+/**
+ * A classe {@link ControllerExceptionHandler} é um manipulador de exceções global para controladores.
+ * Ela lida com as exceções lançadas durante o processamento de solicitações e gera respostas de erro apropriadas.
+ * A classe é anotada com {@link RestControllerAdvice} para aplicar o tratamento de exceção globalmente
+ * a todas as classes de controlador.
+ *
+ * @author diegosneves
+ */
+@RestControllerAdvice
+public class ControllerExceptionHandler {
+
+    /**
+     * Manipula exceções gerais e retorna uma resposta de erro apropriada.
+     *
+     * @param exception A exceção que ocorreu.
+     * @return Uma {@link ResponseEntity} contendo um {@link ExceptionDTO} com a mensagem da exceção e um código de status HTTP
+     */
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ExceptionDTO> generalError(Exception exception) {
+        ExceptionDTO dto = new ExceptionDTO(exception.getMessage(), HttpStatus.BAD_REQUEST.value());
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(dto);
+    }
+
+}
