+++ src/main/java/diegosneves/github/hexagonal/adapters/rest/dto/ExceptionDTO.java
@@ -1,0 +1,4 @@
+package diegosneves.github.hexagonal.adapters.rest.dto;
+
+public record ExceptionDTO(String message, int statusCode) {
+}
