+++ src/main/java/diegosneves/github/hexagonal/adapters/rest/request/ProductRequest.java
@@ -1,0 +1,17 @@
+package diegosneves.github.hexagonal.adapters.rest.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder
+public class ProductRequest {
+    private String productName;
+    private Double productPrice;
+}
