--- src/main/java/diegosneves/github/hexagonal/adapters/service/ProductEntityServiceContract.java
@@ -1,18 +1,0 @@
-package diegosneves.github.hexagonal.adapters.service;
-
-import diegosneves.github.hexagonal.adapters.request.ProductRequest;
-import diegosneves.github.hexagonal.adapters.response.ProductResponse;
-import diegosneves.github.hexagonal.app.domain.product.entity.ProductContract;
-import diegosneves.github.hexagonal.app.exceptions.ProductException;
-
-public interface ProductEntityServiceContract {
-
-    ProductResponse get(String id) throws ProductException;
-
-    ProductResponse create(ProductRequest request);
-
-    ProductResponse enable(String id);
-
-    ProductResponse disable(String id);
-
-}
