+++ src/main/java/diegosneves/github/hexagonal/app/mapper/MapperStrategy.java
@@ -1,0 +1,7 @@
+package diegosneves.github.hexagonal.app.mapper;
+
+public interface MapperStrategy <T, E> {
+
+    T mapper(E origin);
+
+}
